moved lang in prep for more complex lang categorisation
