+++ src/grafzahl/count_modes/custom_modes/language.rs
@@ -4,7 +4,7 @@ use std::fmt::{Display, Formatter};
 use std::ops::Add;
 use crate::AppState;
 use crate::grafzahl::count_modes::countable::Countable;
-use crate::grafzahl::languages::get_lang;
+use crate::grafzahl::language::languages::get_lang;
 
 #[derive(Default, Clone)]
 pub(crate) struct LanguageCount(HashMap<String, u32>);
