+++ src/grafzahl/mod.rs
@@ -1,4 +1,4 @@
-pub mod languages;
+pub mod language;
 pub mod ignore_checker;
 pub mod count_modes;
 mod generic_counter;
